pyproject.toml
[tool.poetry]
name = "effective-git-messages"
version = "0.1.0"
description = ""
authors = ["Fernando Sedrez <fmsedrez@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
google-generativeai = "^0.3.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
